refactor: CustomCheckbox 컴포넌트 적용  및 폴더 구조 정리
